backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.11.2
pydantic-settings==2.8.1
openai==1.72.0
email-validator==2.2.0
PyJWT==2.9.0
pypdf==5.4.0
python-multipart==0.0.9
